+++ flow-typed/def.js
@@ -1,0 +1,4 @@
+type AppTitleProps = {
+  title: string,
+  link?: string
+};