+++ src/App.js
@@ -3,6 +3,8 @@ import React, { Component } from 'react';
 import { Route, Switch } from 'react-router-dom';
 import { Provider } from 'react-redux';
 import { ConnectedRouter } from 'connected-react-router';
+// $FlowFixMe
+import 'antd/dist/antd.less'
 import history from './store/history';
 import store from './store';
 import './styles/reset';
@@ -11,10 +13,9 @@ import { css } from 'react-emotion';
 import NoComponent from './components/NoComponent';
 import Menu from './components/Menu';
 import Header from './components/Header';
-// $FlowFixMe
-import 'antd/dist/antd.less';
+import NotificationList from './components/NotificationList'
 
-const sideColor = '#343434;';
+const sideColor = '#000000;';
 const styles = {
   layout: css`
     display: flex;
@@ -38,19 +39,17 @@ const styles = {
 	`,
   content: css`
     display: block;
-    background: white;
+    background: #f0f2f5;
     flex-grow: 1;
+    max-height: 100vh;
+    overflow: auto;
   `,
   sidemenu: css`
-    display: block;
     flex-grow: 0;
     flex-shrink: 0;
-    width: 243px;
     box-sizing: border-box;
-    padding: 30px 0px 30px 21px;
-    background: white;
-    border-right: 2px solid black;
     background: ${sideColor};
+    z-index: 1;
   `,
 };
 
@@ -71,13 +70,10 @@ export default class App extends Component<any> {
     return (
       <Provider store={store}>
         <div className={styles.layout}>
-          <Header/>
           <div className={styles.main}>
-
-            <div className={styles.sidemenu}>
-              <Menu/>
-            </div>
+            <Menu className={styles.sidemenu}/>
             <div className={styles.content}>
+              <Header/>
               <ConnectedRouter history={history}>
                 <Switch >
                   {mapRoutesModule()}
@@ -85,6 +81,7 @@ export default class App extends Component<any> {
                 </Switch>
               </ConnectedRouter>
             </div>
+            <NotificationList />
           </div>
         </div>
       </Provider>