+++ src/components/Menu/Menu.test.js
@@ -1,20 +1,23 @@
-import {Menu} from './Menu'
+import {Index} from '.'
 import renderer from 'react-test-renderer';
 import * as React from 'react';
-import * as constants from '../store/constants';
+import * as constants from '../../store/constants';
+
+jest.mock('../Scrollbar.js', () => ({ children }) => <div>{children}</div>)
+
 
 describe('Menu', () => {
   const rootPath = '/';
   const testPath = '/test';
   it('should not be called if no menu items', () => {
     const dispatch = jest.fn();
-    const component = renderer.create(<Menu
+    const component = renderer.create(<Index
       path={rootPath}
       menu={[]}
       key={'menu'}
       dispatch={dispatch}
     />);
-    component.update(<Menu
+    component.update(<Index
       path={rootPath}
       menu={[]}
       key={'menu'}
@@ -25,8 +28,8 @@ describe('Menu', () => {
   it('should not be called if no root location and menu', () => {
     const dispatch = jest.fn();
 
-    const component = renderer.create(<Menu path={testPath} menu={[]} key={'menu'} dispatch={dispatch}/>);
-    component.update(<Menu path={testPath} menu={[]} key={'menu'} dispatch={dispatch}/>)
+    const component = renderer.create(<Index path={testPath} menu={[]} key={'menu'} dispatch={dispatch}/>);
+    component.update(<Index path={testPath} menu={[]} key={'menu'} dispatch={dispatch}/>)
     expect(dispatch).not.toBeCalled();
   })
   it('should not be called if no root location', () => {
@@ -41,8 +44,8 @@ describe('Menu', () => {
         items: [],
       },
     ];
-    const component = renderer.create(<Menu path={testPath} menu={otherMenu} key={'menu'} dispatch={dispatch}/>);
-    component.update(<Menu path={testPath} menu={otherMenu} key={'menu'} dispatch={dispatch}/>)
+    const component = renderer.create(<Index path={testPath} menu={otherMenu} key={'menu'} dispatch={dispatch}/>);
+    component.update(<Index path={testPath} menu={otherMenu} key={'menu'} dispatch={dispatch}/>)
     expect(dispatch).not.toBeCalled();
   })
   it('should not be called if no root location and change menu items in progress', () => {
@@ -75,8 +78,8 @@ describe('Menu', () => {
         items: [],
       },
     ];
-    const component = renderer.create(<Menu path={testPath} menu={otherMenu} key={'menu'} dispatch={dispatch}/>);
-    component.update(<Menu path={testPath} menu={testMenu} key={'menu'} dispatch={dispatch}/>)
+    const component = renderer.create(<Index path={testPath} menu={otherMenu} key={'menu'} dispatch={dispatch}/>);
+    component.update(<Index path={testPath} menu={testMenu} key={'menu'} dispatch={dispatch}/>)
     expect(dispatch).not.toBeCalled();
   })
   it('should not be called if has a selected item - single', () => {
@@ -91,8 +94,8 @@ describe('Menu', () => {
         items: [],
       },
     ];
-    const component = renderer.create(<Menu path={rootPath} menu={testMenu} key={'menu'} dispatch={dispatch}/>);
-    component.update(<Menu path={rootPath} menu={testMenu} key={'menu'} dispatch={dispatch}/>);
+    const component = renderer.create(<Index path={rootPath} menu={testMenu} key={'menu'} dispatch={dispatch}/>);
+    component.update(<Index path={rootPath} menu={testMenu} key={'menu'} dispatch={dispatch}/>);
     expect(dispatch).not.toBeCalled();
   })
   it('should not be called if no has a selected item - multi items', () => {
@@ -115,8 +118,8 @@ describe('Menu', () => {
         items: [],
       },
     ];
-    const component = renderer.create(<Menu path={rootPath} menu={testMenu} key={'menu'} dispatch={dispatch}/>);
-    component.update(<Menu path={rootPath} menu={testMenu} key={'menu'} dispatch={dispatch}/>)
+    const component = renderer.create(<Index path={rootPath} menu={testMenu} key={'menu'} dispatch={dispatch}/>);
+    component.update(<Index path={rootPath} menu={testMenu} key={'menu'} dispatch={dispatch}/>)
     expect(dispatch).not.toBeCalled();
   })
   it('should be called if have menu items with no selection - single', () => {
@@ -132,8 +135,8 @@ describe('Menu', () => {
         items: [],
       },
     ];
-    const component = renderer.create(<Menu path={rootPath} menu={testMenu} key={'menu'} dispatch={dispatch}/>);
-    component.update(<Menu path={rootPath} menu={testMenu} key={'menu'} dispatch={dispatch}/>);
+    const component = renderer.create(<Index path={rootPath} menu={testMenu} key={'menu'} dispatch={dispatch}/>);
+    component.update(<Index path={rootPath} menu={testMenu} key={'menu'} dispatch={dispatch}/>);
     expect(dispatch).toBeCalled();
     expect(actions).toHaveLength(1);
     const pushAction = actions[0];
@@ -162,8 +165,8 @@ describe('Menu', () => {
         items: [],
       },
     ];
-    const component = renderer.create(<Menu path={rootPath} menu={testMenu} key={'menu'} dispatch={dispatch}/>);
-    component.update(<Menu path={rootPath} menu={testMenu} key={'menu'} dispatch={dispatch}/>);
+    const component = renderer.create(<Index path={rootPath} menu={testMenu} key={'menu'} dispatch={dispatch}/>);
+    component.update(<Index path={rootPath} menu={testMenu} key={'menu'} dispatch={dispatch}/>);
     expect(dispatch).toBeCalled();
     expect(actions).toHaveLength(1);
     const pushAction = actions[0];