+++ src/index.js
@@ -4,13 +4,19 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 import App from './App'
 import {Router, Route, Switch} from 'react-router-dom'
-import {Button} from 'antd'
 import AppTitle from './components/AppTitle';
 import core from './coreInstance';
+import testSvg from './components/Notification/success-circle.svg'
 
 const rootElement = document.getElementById('root')
 if (rootElement) {
   ReactDOM.render(<App/>, rootElement)
+}
+
+const textElements = [];
+
+for (let i =0; i < 100; i++) {
+  textElements.push(<div style={{marginBottom: '20px'}} key={i}>Test text overflow row: {i}</div>)
 }
 
 class Test extends React.Component<null> {
@@ -18,26 +24,30 @@ class Test extends React.Component<null> {
   render() {
     return (
       <div>
-        <AppTitle title="Root" />
-        test namespace:
+        <AppTitle title="Root" link={'/test/test2'}  />
+        Tarantool - Frontend Core
         <Router history={core.history}>
           <Switch>
             <Route
-              path={'/test/test2'}
+              path={'/test/test'}
               component={() => (
                 <div>
-                  <AppTitle title="Test 2" />
+                  <AppTitle title="Test 2"/>
+                  <AppTitle title="Test 3"/>
+                  <AppTitle title="Test 4"/>
+                  <AppTitle title="Test 6"/>
                   <p>
-                    <Button type={'primary'}>
-                      status
-                    </Button>
+                      Test title page
                   </p>
-                  ыфаафыыаф
                 </div>
               )}
             />
-            <Route path={'/test/test'} component={() => <div><AppTitle title="Test" />1</div>}/>
-            <Route path={'/'} component={() => <div>ЬЬЬ</div>}/>
+            <Route path={'/test/sub'} component={() => <div>
+              <AppTitle title="Test" />
+                {textElements}
+            </div>}/>
+            <Route path={'/test/icon'} component={() => <div>Just content</div>}/>
+            <Route path={'/'} component={() => <div>Some root</div>}/>
           </Switch>
         </Router>
       </div>
@@ -48,31 +58,31 @@ class Test extends React.Component<null> {
 core.register(
   'test',
   [
-    { label: 'Cluster', path: '/test/test2' },
     {
-      label: 'Dashboard',
+      label: 'Simple Title Example',
       path: '/test/test',
-      items: [
-        { label: 'Subitem', path: '/test/test/test' },
-        { label: 'Subitem 2', path: '/test/test/test2' },
-      ],
+      icon: 'hdd',
+    },
+    {
+      label: 'SubItems Example',
+      path: '/test/sub',
+      items: textElements.map((r, i) => ({label: `Subitem ${i}`, path: `/test/sub/${i}`})),
+    },
+    {
+      label: 'Test custom icon',
+      path: '/test/icon',
+      icon: <img src={testSvg} style={{width: 14, height: 14}} />,
     },
   ],
   Test,
   'react'
 )
 
-core.register(
-  'space',
-  [
-    { label: 'Space Cluster', path: '/space/test2' },
-    { label: 'Space Dashboard', path: '/space/test' },
-  ],
-  Test,
-  'react'
-)
 
-core.setHeaderComponent(<div style={{position: 'absolute', right: 0, top: 0}}><span >Absolute</span></div>)
+core.notify({ type: 'success', title: 'Hello', message: `Create the replicaset <b>"Name of replicaset #2"</b> to with server "pim.dmitrov_ storages_ msk_eapi_trn57_5" executed successfully`})
+core.notify({ type: 'error', title: 'hello', message: `Create the replicaset "Name of replicaset #2" to with server "pim.dmitrov_ storages_ msk_eapi_trn57_5" executed successfully`})
+core.notify({ type: 'default', title: 'hello', message: `Create the replicaset "Name of replicaset #2" to with server "pim.dmitrov_ storages_ msk_eapi_trn57_5" executed successfully`})
+core.notify({ type: 'success', title: 'hello', message: `Create the replicaset "Name of replicaset #2" to with server "pim.dmitrov_ storages_ msk_eapi_trn57_5" executed successfully`})
 setTimeout(() => {
   core.setHeaderComponent(
     null