+++ src/store/index.js
@@ -1,34 +1,43 @@
 // @flow
 import { createStore, applyMiddleware, compose, combineReducers } from 'redux';
-import mainReducer, { type MenuState } from './reducer';
-import { changeTitleMiddleware } from './title/middleware';
-import appTitleReducer, { type AppTitleState } from './title/reducer';
+import menuReducer, { type MenuState } from './reducers/menu';
+import notificationsReducer, { type NotificationState } from './reducers/notifications';
+import uiReducer from './reducers/ui';
+import { changeTitleMiddleware } from './middlewares/title';
+import appTitleReducer, { type AppTitleState } from './reducers/title';
 import thunk from 'redux-thunk';
 import { connectRouter, routerMiddleware } from 'connected-react-router';
 import history, { APP_PATH_PREFIX } from './history';
 import CoreInstance from '../coreInstance';
 import * as constants from './constants';
 import { type FSA } from '../core';
+import { startTimer } from './middlewares/notifications';
+import type { UiState } from './reducers/ui'
+import { sendNotification } from './actions/notifications'
 
 export type State = {
   menu: MenuState,
-  appTitle: AppTitleState
+  appTitle: AppTitleState,
+  notifications: NotificationState,
+  ui: UiState,
 }
 
 const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
 
 const store = createStore(
   connectRouter(history)(combineReducers({
-    menu: mainReducer.reduce,
+    menu: menuReducer.reduce,
     appTitle: appTitleReducer,
+    ui: uiReducer,
+    notifications: notificationsReducer,
   })),
-  composeEnhancers(applyMiddleware(thunk, routerMiddleware(history), mainReducer.middleware, changeTitleMiddleware))
+  composeEnhancers(applyMiddleware(thunk, routerMiddleware(history), menuReducer.middleware, changeTitleMiddleware))
 );
 
 CoreInstance.subscribe('registerModule', () => {
   const modules = CoreInstance.getModules();
   for (const module of modules) {
-    const added = mainReducer.processModule(module);
+    const added = menuReducer.processModule(module);
     if (added) {
       store.dispatch({
         type: constants.RESET,
@@ -42,5 +51,7 @@ CoreInstance.subscribe('registerModule', () => {
 });
 
 CoreInstance.subscribe('dispatchToken', (token: FSA) => store.dispatch(token));
+
+startTimer(store)
 
 export default store;