+++ src/store/middlewares/notifications.js
@@ -1,0 +1,7 @@
+import { checkNotifications } from '../actions/notifications'
+
+export const startTimer = (store) => {
+  setInterval(() => {
+    store.dispatch(checkNotifications())
+  }, 1000)
+}