+++ CHANGELOG.md
@@ -4,10 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [6.0.0]
 
 - Add title in app header and sets window title.
 - Fix side menu bug because of it subitems couldn't be rendered as selected.
+- Change behaviour of menu expand
+- Redesign UI
+- Add notify method for notification
+- Deprecate usage of `setHeaderComponent`
 
 ## [5.0.2] - 2019-06-25